pyproject.toml
[project]
name = "dspy-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-devops>=7.1.0b4",
    "dspy==3.0.3",
    "fastmcp==2.13.0.2",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
